backend/requirements.txt

flask
python-dotenv
flask_cors
aiohttp>=3.8.1,<4.0.0
azure-functions
azure-identity
azure-mgmt-core
azure-mgmt-resourcegraph
azure-mgmt-subscription
msrestazure
azure-data-tables
azure-mgmt-resource
azure-storage-blob
azure.mgmt.support
flask[async]
azure-messaging-webpubsubservice
azure-data-tables
flask-caching